Refactor cachingConfigMapManager
